Move SplitIssueTree to splittree package (#14)
